Refactor middleware architecture with separated concerns and barrel exports
Break out monolithic system-context.ts into focused middleware files:

- system-context.ts: Core system initialization and error handling only
- response-json.ts: JSON response formatting middleware
- response-yaml.ts: YAML response middleware (for meta API)
- response-file.ts: File response middleware (future functionality)
- index.ts: Barrel export for clean imports

Update all route imports to use barrel export pattern:
- All 24+ route files now import from @src/lib/middleware/index.js
- Main index.ts uses single import for all middleware
- Consistent import patterns across entire codebase

Benefits:
- Single responsibility: Each middleware file has one clear purpose
- Better maintainability: Easier to modify individual middleware
- Cleaner imports: Barrel export eliminates import clutter
- Modular architecture: Middleware can be developed independently
