+++ src/lib/middleware/index.ts
@@ -1,0 +1,14 @@
+/**
+ * Middleware Barrel Export
+ * 
+ * Clean middleware organization for Hono route handling:
+ * - System context and error handling
+ * - Response formatting (JSON, YAML, File)
+ * - Development and security middlewares
+ */
+
+export { systemContextMiddleware, setRouteResult } from './system-context.js';
+export { responseJsonMiddleware } from './response-json.js';
+export { responseYamlMiddleware } from './response-yaml.js';
+export { responseFileMiddleware } from './response-file.js';
+export { localhostDevelopmentOnlyMiddleware } from './localhost-development-only.js';
