+++ src/lib/middleware/response-file.ts
@@ -1,0 +1,19 @@
+/**
+ * File Response Middleware
+ * 
+ * Placeholder for file upload/download response formatting
+ */
+
+import type { Context, Next } from 'hono';
+
+/**
+ * File response middleware for future /api/file/* routes
+ * 
+ * Placeholder for file upload/download response formatting
+ */
+export async function responseFileMiddleware(context: Context, next: Next) {
+    await next();
+    
+    // Future: Handle file responses, content-type headers, streaming, etc.
+    // Could format file metadata, handle multipart responses, etc.
+}
